requirements.txt
asgiref==3.2.7
autopep8==1.5.4
boto3==1.12.44
botocore==1.15.44
dj-database-url==0.5.0
Django==3.0.5
django-admin-sortable2==0.7.7
django-braces==1.14.0
django-extensions==3.0.8
django-heroku==0.3.1
django-model-utils==4.0.0
django-star-ratings==0.9.1
django-storages==1.9.1
django-tinymce==3.0.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.0.1
